Change foreign key field type to big int for draft budgets

+++ db/migrate/20210301141650_add_template_to_draft_budget.rb
@@ -1,5 +1,5 @@
 class AddTemplateToDraftBudget < ActiveRecord::Migration[5.2]
   def change
-      add_reference :draft_budgets, :funding_template, foreign_key: true, type: :integer
+      add_reference :draft_budgets, :funding_template, foreign_key: true, type: :bigint
   end
 end
